Cleaned up carousel elements - added custom observers to prevent cycling on media playback - added intersect observer to pause carousel and prevent layout shifts (on mobile)
