--- apps/verification-portal/src/lib/components/carousel/CarouselSlide.ts
@@ -1,22 +1,0 @@
-// https://github.com/themesberg/flowbite-svelte/blob/main/src/lib/carousel/CarouselSlide.ts
-
-export const canChangeSlide = ({
-  lastSlideChange,
-  slideDuration,
-  slideDurationRatio = 1,
-}: {
-  lastSlideChange: Date;
-  slideDuration: number;
-  slideDurationRatio?: number; // Allows for starting a new transition before the previous completes
-}) => {
-  if (
-    lastSlideChange &&
-    new Date().getTime() - lastSlideChange.getTime() <
-      slideDuration * slideDurationRatio
-  ) {
-    console.warn("Can't change slide yet, too soon");
-    return false;
-  }
-
-  return true;
-};
